requirements.txt
# requirements.txt - All versions pinned for consistency
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0

# Data processing
pandas==2.1.3
numpy==1.24.3
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Embeddings and NLP - Note: These will download models once and cache them
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0

# Analysis
scikit-learn==1.3.2
scipy==1.11.4
umap-learn==0.5.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
wordcloud==1.9.2

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
joblib==1.3.2
python-dotenv==1.0.0
openpyxl==3.1.2  # For Excel export